Fix: avoid warning on DEC Alpha's (the test main() pgm will probably not work there)

+++ termcap/termcap.c
@@ -498,7 +498,7 @@ tgetent (bp, name)
  ret:
   term_entry = bp;
   if (malloc_size)
-    return (int) bp;
+    return (int)(long)bp;
   return 1;
 }
